+++ src/frontend/Orion/Organizations/Obm/TemplateTable/TemplateTable.js
@@ -1,4 +1,5 @@
 import React, { useState } from 'react'
+import _ from 'lodash'
 
 import { useTable, useBlockLayout, useFilters, useSortBy } from 'react-table'
 import { useSticky } from 'react-table-sticky'
@@ -8,6 +9,9 @@ import Headers from './Headers'
 import Rows from './Rows'
 import ModalManager from './ModalManager'
 import sortTypes from './custom-sort-types'
+
+import ExportExcelButton from 'frontend/components/ExportExcelButton'
+import formatDataForExport from './formatDataForExport'
 
 const DefaultColumnFilter = ({ column: { filterValue, preFilteredRows, setFilter } }) => {
   const count = preFilteredRows.length
@@ -27,10 +31,10 @@ const DEFAULT_COLUMN = {
   Filter: DefaultColumnFilter,
 }
 
-const TemplateTable = ({ columns, data, modalColMap }) => {
+const TemplateTable = ({ columns, data, modalColMap, exportProps }) => {
   const [modalCell, setModalCell] = useState(null)
 
-  const { getTableProps, getTableBodyProps, headerGroups, rows, prepareRow } = useTable(
+  const { getTableProps, getTableBodyProps, headerGroups, rows, prepareRow, sortedRows } = useTable(
     {
       columns,
       data,
@@ -45,8 +49,20 @@ const TemplateTable = ({ columns, data, modalColMap }) => {
     useSticky
   )
 
+  const dataFormattedForExport = formatDataForExport(sortedRows, columns)
+
   return (
     <>
+      <div style={{ margin: '0 24px 0 auto' }}>
+        <ExportExcelButton
+          data={dataFormattedForExport}
+          // buttonStyle={} // if custom button styling is needed, style here
+          {...exportProps}
+        >
+          Export Current View to Excel
+        </ExportExcelButton>
+      </div>
+
       <TableWrapper>
         <div className="table sticky" style={{ height: '100%' }} {...getTableProps()}>
           <Headers headerGroups={headerGroups} />