+++ src/frontend/components/ExportExcelButton/ExportExcelButton.js
@@ -1,3 +1,5 @@
+/** @jsx jsx */
+import { jsx } from '@emotion/core'
 import React from 'react'
 import PropTypes from 'prop-types'
 import styled from '@emotion/styled'
@@ -7,38 +9,34 @@ import { Colors } from '../../utils/pulseStyles'
 
 const primaryColor = Colors.PRIMARY
 
-const ExportButton = styled.button({
-  background: transparentize(0.85,primaryColor),
-  color: primaryColor,
-  fontSize: 12,
-  fontWeight: 700,
-  padding: '8px 12px',
-  borderRadius: 4,
-  cursor: 'pointer',
-  margin: '12px 0px',
-  ':hover': {
-    background: transparentize(0.65, primaryColor),
+const ExportButton = styled.button(
+  {
+    background: transparentize(0.85, primaryColor),
+    color: primaryColor,
+    fontSize: 12,
+    fontWeight: 700,
+    padding: '8px 12px',
+    borderRadius: 4,
+    cursor: 'pointer',
+    margin: '12px 0px',
+    ':hover': {
+      background: transparentize(0.65, primaryColor),
+    },
+  },
+  ({ disabled }) => {
+    return disabled
+      ? {
+          background: transparentize(0.85, Colors.MEDIUM_GRAY_2),
+          color: 'grey',
+          cursor: 'not-allowed',
+        }
+      : {}
   }
-}, ({ disabled }) => {
-  return disabled
-    ? {
-      background: transparentize(0.85, Colors.MEDIUM_GRAY_2),
-      color: 'grey',
-      cursor: 'not-allowed',
-    }
-    : {}
-})
+)
 
-const ExportExcelButton = ({
-  isDisabled,
-  onClick,
-  children,
-}) => {
+const ExportExcelButton = ({ isDisabled, onClick, children, buttonStyle }) => {
   return (
-    <ExportButton
-      disabled={isDisabled}
-      onClick={onClick}
-    >
+    <ExportButton css={buttonStyle} disabled={isDisabled} onClick={onClick}>
       {children}
     </ExportButton>
   )