+++ CHANGELOG.md
@@ -14,6 +14,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - Started: 6-30-20
 
+### Added
+- Improve Obm Template Table
+  - Add ability to set sticky columns
+  - Refactor table components
+  - Add export to excel functionality
+
 ## [2.10.2] - Started: 6-30-20
 
 ### Changed